Separating HeaderFilterMessage component into it's own file
