+++ client/app/components/reader/DocumentListHeader.jsx
@@ -2,60 +2,11 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import SearchBar from '../SearchBar';
 import { connect } from 'react-redux';
-import { setSearch, clearAllFilters, clearSearch } from '../../reader/actions';
+import { setSearch, clearSearch } from '../../reader/actions';
 import _ from 'lodash';
-import Analytics from '../../util/AnalyticsUtil';
 import ApiUtil from '../../util/ApiUtil';
 import DocumentsCommentsButton from '../../reader/DocumentsCommentsButton';
-
-class HeaderFilterMessage extends React.PureComponent {
-  render() {
-    const props = this.props;
-
-    const categoryFilters = Object.keys(props.docFilterCriteria.category).some((category) =>
-      props.docFilterCriteria.category[category]
-    );
-    const tagFilters = Object.keys(props.docFilterCriteria.tag).some((tag) =>
-      props.docFilterCriteria.tag[tag]
-    );
-
-    const categoryCount = _(props.docFilterCriteria.category).
-      values().
-      compact().
-      size();
-    const tagCount = _.size(props.docFilterCriteria.tag);
-
-    const filteredCategories = _.compact([
-      categoryFilters && `Categories (${categoryCount})`,
-      tagFilters && `Issue tags (${tagCount})`,
-      props.viewingDocumentsOrComments === 'comments' && 'Comments'
-    ]).join(', ');
-
-    if (!filteredCategories.length) {
-      return null;
-    }
-
-    return <p className="document-list-filter-message">Filtering by: {filteredCategories}. <a
-      href="#"
-      id="clear-filters"
-      onClick={props.clearAllFilters}>
-      Clear all filters.</a></p>;
-  }
-}
-
-const headerFilterMessageMapDispatchToProps = (dispatch) => ({
-  clearAllFilters: () => {
-    Analytics.event('Claims Folder', 'click', 'Clear all filters');
-    dispatch(clearAllFilters());
-  }
-});
-
-const headerFilterMessageMapStateToProps = (state) => ({
-  viewingDocumentsOrComments: state.viewingDocumentsOrComments
-});
-
-const ConnectedHeaderFilterMessage = connect(headerFilterMessageMapStateToProps,
-                                             headerFilterMessageMapDispatchToProps)(HeaderFilterMessage);
+import HeaderFilterMessage from './HeaderFilterMessage';
 
 class DocumentListHeader extends React.Component {
   // Record the search value for analytics purposes, don't worry if it fails.
@@ -88,7 +39,7 @@ class DocumentListHeader extends React.Component {
         </div>
         <DocumentsCommentsButton />
       </div>
-        <ConnectedHeaderFilterMessage docFilterCriteria={props.docFilterCriteria} />
+        <HeaderFilterMessage docFilterCriteria={props.docFilterCriteria} />
     </div>;
   }
 }
